+++ packages/sdk/src/models/ion-gateway-options.ts
@@ -1,6 +1,7 @@
 import { IStorage } from 'src/storage/models/storage.interface';
 import { EventDispatcher } from 'src/tracker/event-dispatcher';
 import { SdkActionEvent } from 'src/tracker/types';
+import { RequiredFeatures } from './wallet';
 
 /**
  * TonConnect constructor options
@@ -34,6 +35,11 @@ export interface TonConnectOptions {
      * @default Infinity
      */
     walletsListCacheTTLMs?: number;
+
+    /**
+     * Required features for wallets. If wallet doesn't support required features, it will be disabled.
+     */
+    walletsRequiredFeatures?: RequiredFeatures;
 
     /**
      * Allows to disable auto pause/unpause SSE connection on 'document.visibilitychange' event. It is not recommended to change default behaviour.