+++ packages/sdk/src/models/wallet/index.ts
@@ -19,3 +19,4 @@ export {
     isWalletInfoRemote,
     isWalletInfoInjected
 } from './wallet-info';
+export { RequiredFeatures, RequiredSendTransactionFeature } from './require-feature';