+++ packages/sdk/src/models/wallet/require-feature.ts
@@ -1,0 +1,24 @@
+/**
+ * Required features for wallets.
+ */
+export type RequiredFeatures = {
+    /**
+     * Required features for the send transaction feature.
+     */
+    sendTransaction?: RequiredSendTransactionFeature;
+};
+
+/**
+ * Required features for the send transaction feature.
+ */
+export type RequiredSendTransactionFeature = {
+    /**
+     * Minimum number of messages to send.
+     */
+    minMessages?: number;
+
+    /**
+     * Whether extra currency is required.
+     */
+    extraCurrencyRequired?: boolean;
+};