+++ packages/ui/src/app/models/personalized-wallet-info.ts
@@ -1,5 +1,0 @@
-import { WalletInfo } from '@ion-gateway/sdk';
-
-export type PersonalizedWalletInfo = WalletInfo & {
-    isPreferred?: boolean;
-};