+++ packages/ui/src/app/models/ui-wallet-info.ts
@@ -1,0 +1,6 @@
+import { WalletInfo } from '@ion-gateway/sdk';
+
+export type UIWalletInfo = WalletInfo & {
+    isPreferred?: boolean;
+    isSupportRequiredFeatures: boolean;
+};