+++ packages/ui/src/app/state/app.state.ts
@@ -1,4 +1,4 @@
-import { ConnectAdditionalRequest, ITonConnect } from '@ion-gateway/sdk';
+import { ConnectAdditionalRequest, RequiredFeatures, ITonConnect } from '@ion-gateway/sdk';
 import { createStore } from 'solid-js/store';
 import { Locales } from 'src/models/locales';
 import { WalletsListConfiguration } from 'src/models/wallets-list-configuration';
@@ -15,6 +15,7 @@ export type AppState = {
     twaReturnUrl: `${string}://${string}` | undefined;
     preferredWalletAppName: string | undefined;
     enableAndroidBackHandler: boolean;
+    walletsRequiredFeatures: RequiredFeatures | undefined;
 };
 
 export const [appState, setAppState] = createStore<AppState>({