+++ packages/ui/src/ion-gateway-ui.ts
@@ -1,6 +1,7 @@
 import type {
     Account,
     ConnectAdditionalRequest,
+    RequiredFeatures,
     WalletInfoCurrentlyEmbedded
 } from '@ion-gateway/sdk';
 import {
@@ -19,7 +20,7 @@ import { TonConnectUIError } from 'src/errors/ion-gateway-ui.error';
 import { TonConnectUiCreateOptions } from 'src/models/ion-gateway-ui-create-options';
 import { PreferredWalletStorage, WalletInfoStorage } from 'src/storage';
 import {
-    createMacrotask, createMacrotaskAsync,
+    createMacrotaskAsync,
     getSystemTheme,
     preloadImages,
     subscribeToThemeChange
@@ -67,6 +68,8 @@ export class TonConnectUI {
     private actionsConfiguration?: ActionConfiguration;
 
     private readonly walletsList: Promise<WalletInfo[]>;
+
+    public readonly walletsRequiredFeatures?: RequiredFeatures;
 
     private connectRequestParametersCallback?: (
         parameters: ConnectAdditionalRequest | undefined
@@ -195,7 +198,8 @@ export class TonConnectUI {
         } else if (options && 'manifestUrl' in options && options.manifestUrl) {
             this.connector = new TonConnect({
                 manifestUrl: options.manifestUrl,
-                eventDispatcher: options?.eventDispatcher
+                eventDispatcher: options.eventDispatcher,
+                walletsRequiredFeatures: options.walletsRequiredFeatures
             });
         } else {
             throw new TonConnectUIError(
@@ -231,6 +235,8 @@ export class TonConnectUI {
         this.transactionModal = new TransactionModalManager({
             connector: this.connector
         });
+
+        this.walletsRequiredFeatures = options.walletsRequiredFeatures;
 
         this.walletsList = this.getWallets();
 
@@ -421,7 +427,7 @@ export class TonConnectUI {
             sendExpand();
         }
 
-        const { notifications, modals, returnStrategy, twaReturnUrl, skipRedirectToWallet } =
+        const { notifications, modals, returnStrategy, twaReturnUrl } =
             this.getModalsAndNotificationsConfiguration(options);
 
         widgetController.setAction({