+++ packages/ui/src/library.ts
@@ -1,7 +1,5 @@
 export { TonConnectUI as default } from './ion-gateway-ui';
 export { TonConnectUI } from './ion-gateway-ui';
-export type {
-    UserActionEvent,
-} from './tracker/types';
+export type { UserActionEvent } from './tracker/types';
 export * from './models';
 export * from './errors';