+++ packages/ui/src/models/ion-gateway-ui-options.ts
@@ -2,6 +2,7 @@ import { Locales } from 'src/models/locales';
 import { UIPreferences } from 'src/models/ui-preferences';
 import { WalletsListConfiguration } from 'src/models/wallets-list-configuration';
 import { ActionConfiguration } from 'src/models/action-configuration';
+import { RequiredFeatures } from '@ion-gateway/sdk';
 
 export interface TonConnectUiOptions {
     /**
@@ -25,6 +26,11 @@ export interface TonConnectUiOptions {
      * Configuration for the wallets list in the connect wallet modal.
      */
     walletsListConfiguration?: WalletsListConfiguration;
+
+    /**
+     * Required features for wallets. If wallet doesn't support required features, it will be disabled.
+     */
+    walletsRequiredFeatures?: RequiredFeatures;
 
     /**
      * Configuration for action-period (e.g. sendTransaction) UI elements: modals and notifications and wallet behaviour (return strategy).